+++ nvcc_gpu/Neuron.h.cpp
@@ -6,10 +6,10 @@
 class Neuron {
 private:
   int numInputs;
-  std::vector<double>* weights;
-  double delta;
-  double activation;
-  double value;
+  std::vector<float>* weights;
+  float delta;
+  float activation;
+  float value;
 
 public:
   Neuron() {}
@@ -17,12 +17,12 @@ public:
   Neuron(int inputs) {
     numInputs = inputs;
     // There is an extra weight for the bias input.
-    weights = new std::vector<double>(numInputs + 1);
+    weights = new std::vector<float>(numInputs + 1);
 
     // Setup weights with an initial random value between -1 and 1. There is
     // one weight for each input and an additional bias weight.
     for (int i = 0; i < weights->size(); i++) {
-      (*weights)[i] = 10 * (((double)rand() / (double)RAND_MAX) * 2 - 1);
+      (*weights)[i] = 10 * (((float)rand() / (float)RAND_MAX) * 2 - 1);
     }
   }
 
@@ -31,42 +31,42 @@ public:
   }
 
   // Get the corresponding weight.
-  double getWeight(int n) const {
+  float getWeight(int n) const {
     return (*weights)[n];
   }
 
   // Add an update value to a specified input weight.
-  void updateWeight(int pos, double update) {
+  void updateWeight(int pos, float update) {
     (*weights)[pos] += update;
   }
 
   // Get the linear combination of inputs to the neuron.
-  double getActivation() const {
+  float getActivation() const {
     return activation;
   }
 
   // Get the value of the neuron (sigmoid applied to the activation).
-  double getValue() const {
+  float getValue() const {
     return value;
   }
 
   // Set the value of the neuron.
-  void setValue(double v) {
+  void setValue(float v) {
     value = v;
   }
 
   // Get the delta value for this neuron.
-  double getDelta() const {
+  float getDelta() const {
     return delta;
   }
 
   // Set the delta value for this neuron.
-  void setDelta(double new_delta) {
+  void setDelta(float new_delta) {
     delta = new_delta;
   }
 
   // Compute and set the linear combination of inputs to the neuron.
-  void setActivation(double a) {
+  void setActivation(float a) {
     activation = a;
   }
